Use conventional layout for multi-module tests
