+++ tests/change_configuration.rs
